Fix llListSort() for unstrided lists.
llListSort() has issues when the list is unstrided.  Rather than based
on all of the elements of the list it collects the list as a single
element.  As a result the list never gets sorted at all.

NOTE: This does not make llListSort() fully accurate, as noted in the
online documentation, sorting a mixed type list will have undocumented
behaviors.