Fix optimization errors with strings.
If a string function produces constant output that contains LSL
unencodable characters that can be expressed in Haskell strings,
the optimizer will collapse these into invalid string.

The easiest way to see this is that llChar(9) != "\t" (it is actually
"    ".  The only value that can be encoded from the control space
is "\n".  As such we prevent optimization.

NOTE: If the user encodes things themselves they may still be
improperly optimize and interepted by LSLForge.