Eclipse cannot locate source in many cases when using the LSL Debugger.
When debugging there are many cases where the source won't be found.
This change adjusts the way the file is looked up.  Previously it was
being looked up using the file name as a URI, this is generally not
going to work out.  Now we convert the absolute file name to a URI first
and lookup using this.