Big update to modernize Haskell and support JSON functions.
Made a lot of small changes to bring Haskell up to date.  This was partially
needed to get the JSON library I wanted to use in and have access to a
few important language features.

In addition I added support for simulating JSON functions (llJson2List,
llJsonGetValue, llJsonSetValue, llJsonValueType, llList2Json), and
llReplaceSubString.